Redesign create page - remove old and necessary packages #1 - redesign create page #8 - rewrite components using vanilla html and js - refactor code and move styling to css files #6
