+++ package.json
@@ -12,21 +12,13 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
     "firebase": "^8.10.0",
-    "fs": "^0.0.1-security",
     "install-custom-font": "^1.1.1",
     "next": "11.1.0",
-    "path": "^0.12.7",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-firebase-hooks": "^3.0.4",
-    "react-particles-js": "^3.5.3",
-    "react-router-dom": "^5.2.0",
-    "react-tsparticles": "^1.33.3",
-    "save-svg-as-png": "^1.4.17",
-    "sharp": "^0.29.0"
+    "save-svg-as-png": "^1.4.17"
   },
   "devDependencies": {
     "eslint": "7.32.0",
