+++ styles/Home.module.css
@@ -6,6 +6,8 @@
   justify-content: center;
   align-items: center;
   min-height: 100vh;
+
+  width: 100%;
 }
 
 .main {
@@ -19,7 +21,7 @@
 
 .footer {
   width: 100%;
-  height: 120px;
+  height: 80px;
   border-top: 1px solid #eaeaea;
   display: flex;
   justify-content: center;
