+++ styles/globals.css
@@ -6,6 +6,22 @@ body {
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 }
 
+.container {
+  height: 100vh;
+  padding: 0.5rem;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  background-color: #3c3c3c;
+}
+
+.container-item-full-height {
+  flex-grow: 1;
+}
+
 a {
   color: inherit;
   text-decoration: none;
@@ -13,15 +29,6 @@ a {
 
 * {
   box-sizing: border-box;
-}
-
-
-#tsparticles {
-  position:fixed !important;
-  left:0;
-  top:0;
-  width:100%;
-  height:100%;
 }
 
 @font-face {
